inconsistency fix

+++ gRPCGraphQLWebSockets/Services/Intefaces/IMessagesService.cs
@@ -4,7 +4,7 @@ namespace gRPCGraphQLWebSockets.Services.Intefaces
 {
     public interface IMessagesService
     {
-        public Message GetMessage(int messageId);
+        public Message GetMessage(long messageId);
 
         public long AddMessage(NewMessage newMessage);
     }
